Adds cool logo to navbar, destroys the previous unprofessional login and instead replaces it with a more angelic login form
