+++ src/Components/Login/Login.css
@@ -10,12 +10,6 @@
     justify-content: center;
 }
 
-.img {
-    height: 100vh;
-    width: 60vw;
-    float: right;
-}
-
 .button-container {
     display: flex;
     flex-direction: column;
