+++ src/Components/Login/Login.js
@@ -1,39 +1,68 @@
-import React from 'react';
-import condo from '../../assets/condo.jpg'
 import { Button } from '@material-ui/core';
-import { Link } from 'react-router-dom';
-import './Login.css';
+import React, { useState, useContext } from 'react';
+import TextField from '@material-ui/core/TextField';
+import condo from '../../assets/condo.jpg'
+import './LoginForm.css';
+import { Context as AuthenticationContext } from '../../context/AuthenticationContext';
 
-function Login () {
+function LoginFormWithContext (props) {
+    const { signin } = useContext(AuthenticationContext);
+    // const { isLoading, error } = state;
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const [error, setError] = useState("");
+
+
+    /**
+     * Handles clicking on the 'sign in' button
+     * @param {*} event 
+     */
+    const handleClick = async (e) => {
+        const data = await signin({ email: email, password: password });
+        if (data) {
+            setError(false);
+            props.history.push('/condo-association');
+        }
+        else {
+            setError(true);
+        }
+    }
+
     return (
-        <div className="main-container" >
-            <img className="img" src={condo} alt="condo"></img>
+        <div className="login-container">
+          <div className="login-form-container">
+            <h1 className="login-title">Login</h1>
 
-            <div className="container2">
-                <h1 className="title">CON System</h1>
+            <form className="form">
+                <TextField
+                    id="outlined-password-input"
+                    label="Email"
+                    type="email"
+                    value={email}
+                    onChange={e => setEmail(e.target.value)}
+                />
+                <TextField className="input"
+                    id="outlined-password-input"
+                    label="Password"
+                    type="password"
+                    autoComplete="current-password"
+                    value={password}
+                    onChange={e => setPassword(e.target.value)}
+                />
+                <div>
+                    <Button
+                        className="signin-btn"
+                        variant="contained"
+                        color="secondary" container
+                        onClick={handleClick}
+                    >Sign in</Button>
+                </div>
 
-            </div>
-
-            <div className="button-container">
-                <Button
-                    variant="contained"
-                    color="secondary" container
-                    style={{ size: 'large', backgroundColor: '#32a895', height: '55px', width: '200px' }}
-                    component={Link} to='/login/loginform'
-                >Sign as admin</Button>
-
-                <Button
-                    variant="contained"
-                    color="secondary" container
-                    style={{ size: 'large', backgroundColor: '#32a895', color: 'white', height: '55px', width: '200px' }}
-                    component={Link} to='/login/loginform'
-                >Sign in as member</Button>
-
-            </div>
-
-
+                {error && <p className="is-error">Your email or password is invalid. Please try again.</p>}
+            </form>
+          </div>
         </div>
     );
 }
 
-export default Login;
+export default LoginFormWithContext;
