--- src/Components/Login/LoginForm.css
@@ -1,40 +1,0 @@
-
-@import url('https://fonts.googleapis.com/css2?family=Arimo&family=Yantramanav:wght@100;300&display=swap');
-.container {
-    height: 100vh;
-}
-
-.container2 {
-    display: flex;
-    min-height: 30vh;
-    align-items: center;
-    justify-content: center;
-}
-
-.img {
-    height: 100vh;
-    width: 60vw;
-    float: right;
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
-    min-height: 30vh;
-}
-
-.title {
-    text-align: center;
-    color: #32a895;
-    font-family: 'Yantramanav', sans-serif;
-}
-
-.is-error {
-    color: red;
-}
-
-.is-success {
-    color: green;
-}
