+++ src/Components/NavBar/NavBar.js
@@ -1,32 +1,14 @@
 import React, { useContext } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
+import condoLogo from '../../assets/logo.png';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
 import { Button } from '@material-ui/core';
 import { Link } from 'react-router-dom';
 import { withRouter } from 'react-router-dom';
 
 import { Context as AuthenticationContext } from '../../context/AuthenticationContext';
 
-const useStyles = makeStyles((theme) => ({
-    root: {
-        flexGrow: 1,
-    },
-    menuButton: {
-        marginRight: theme.spacing(2),
-    },
-    title: {
-        flexGrow: 1,
-    },
-    test: {
-        paddingRight: 10,
-        textTransform: 'none'
-    }
-}));
-
 function NavBar (props) {
-    const classes = useStyles();
     const { signout } = useContext(AuthenticationContext);
 
     const AdminNavBar = () => {
@@ -35,20 +17,17 @@ function NavBar (props) {
             signout();
         }
         return (
-            <div className={classes.root}>
+            <div>
                 <AppBar position="static">
                     <Toolbar style={{ backgroundColor: '#32a895' }}>
-                        <Typography variant="h6" className={classes.title}>
-                            CONSys
-                    </Typography>
+                    <Link className="logo-wrapper" to='/'><img className="logo" src={condoLogo} alt="condo"></img></Link>
                         <div style={{ display: 'flex' }}>
-                            <Button component={Link} to='/admin' color="inherit" className={classes.test} >Admin</Button>
-                            <Button component={Link} to='/condo-association' color="inherit" className={classes.test}>Condo Association</Button>
-                            <Button component={Link} to='/group-home' color="inherit" className={classes.test} >Groups</Button>
-                            <Button component={Link} to='/email' color="inherit" className={classes.test} >Email</Button>
-                            <Button component={Link} to='/my-account' color="inherit" className={classes.test} >My account</Button>
-                            <Button color="inherit" className={classes.test} onClick={handleLogout}>Logout</Button>
-
+                            <Button component={Link} to='/admin' color="inherit" className="nav-link" >Admin</Button>
+                            <Button component={Link} to='/condo-association' color="inherit" className="nav-link">Condo Association</Button>
+                            <Button component={Link} to='/group-home' color="inherit" className="nav-link" >Groups</Button>
+                            <Button component={Link} to='/email' color="inherit" className="nav-link" >Email</Button>
+                            <Button component={Link} to='/my-account' color="inherit" className="nav-link" >My account</Button>
+                            <Button color="inherit" className="nav-link" onClick={handleLogout}>Logout</Button>
                         </div>
                     </Toolbar>
                 </AppBar>
@@ -63,18 +42,16 @@ function NavBar (props) {
             signout();
         }
         return (
-            <div className={classes.root}>
+            <div>
                 <AppBar position="static">
                     <Toolbar style={{ backgroundColor: '#32a895' }}>
-                        <Typography variant="h6" className={classes.title}>
-                            CONSys
-                    </Typography>
+                    <Link className="logo-wrapper" to='/'><img className="logo" src={condoLogo} alt="condo"></img></Link>
                         <div style={{ display: 'flex' }}>
-                            <Button component={Link} to='/condo-association' color="inherit" className={classes.test}>Condo Association</Button>
-                            <Button component={Link} to='/group-home' color="inherit" className={classes.test} >Groups</Button>
-                            <Button component={Link} to='/email' color="inherit" className={classes.test} >Email</Button>
-                            <Button component={Link} to='/my-account' color="inherit" className={classes.test} >My account</Button>
-                            <Button color="inherit" className={classes.test} onClick={handleLogout}>Logout</Button>
+                            <Button component={Link} to='/condo-association' color="inherit" className="nav-link">Condo Association</Button>
+                            <Button component={Link} to='/group-home' color="inherit" className="nav-link" >Groups</Button>
+                            <Button component={Link} to='/email' color="inherit" className="nav-link" >Email</Button>
+                            <Button component={Link} to='/my-account' color="inherit" className="nav-link" >My account</Button>
+                            <Button color="inherit" className="nav-link" onClick={handleLogout}>Logout</Button>
 
                         </div>
                     </Toolbar>
@@ -86,16 +63,14 @@ function NavBar (props) {
 
     const PublicNavBar = () => {
         return (
-            <div className={classes.root}>
+            <div>
                 <AppBar position="static">
                     <Toolbar style={{ backgroundColor: '#32a895' }}>
-                        <Typography variant="h6" className={classes.title}>
-                            CONSys
-                    </Typography>
+                    <Link className="logo-wrapper" to='/'><img className="logo" src={condoLogo} alt="condo"></img></Link>
                         <div style={{ display: 'flex' }}>
-                            <Button component={Link} to='/' color="inherit" className={classes.test} >Ads</Button>
-                            <Button component={Link} to='/login' color="inherit" className={classes.test}>Condo Association</Button>
-                            <Button color="inherit" className={classes.test} component={Link} to='login'>Login</Button>
+                            <Button component={Link} to='/' color="inherit" className="nav-link" >Ads</Button>
+                            <Button component={Link} to='/login' color="inherit" className="nav-link">Condo Association</Button>
+                            <Button color="inherit" className="nav-link" component={Link} to='login'>Login</Button>
                         </div>
                     </Toolbar>
                 </AppBar>
