+++ src/css/GlobalStyles.css
@@ -1,3 +1,72 @@
+:root {
+  --header-height: 101px;
+}
+
+.MuiInput-underline::after {
+  border-color: white !important;
+}
+
+.MuiFormLabel-root {
+  color: white !important;
+}
+
+.MuiInputBase-input {
+  font-size: 16px !important;
+}
+
+.logo {
+  width: 85px;
+}
+
+.logo-wrapper {
+  margin-right: auto;
+}
+
+.nav-link {
+  font-size: 15px !important;
+  font-weight: 600 !important;
+  text-transform: none !important;
+}
+
+.login-container {
+  display: flex;
+  width: 100%;
+  height: calc(100vh - var(--header-height));
+  background: url("/src/assets/condo.jpg") no-repeat;
+  justify-content: center;
+  align-items: center;
+}
+
+.login-title {
+  color: white;
+  text-align: center;
+  font-size: 30px;
+}
+
+.signin-btn {
+  background-color: white !important;
+  color: black !important;
+}
+
+.login-form-container {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  background: #57958b;
+  max-width: 300px;
+  height: 400px;
+  align-items: center;
+  opacity: 1;
+  border-radius: 5px;
+  border: 1px solid white;
+  box-shadow: 0px 0px 7px white;
+}
+
+.login-image-container {
+  width: 100%;
+}
+
 .breadcrumb {
   margin: 15px 0;
 }
